+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 
 ### Chore & Maintenance
 
+- `[expect]` Export default matchers ([#11932](https://github.com/facebook/jest/pull/11932))
+
 ### Performance
 
 ## 27.2.4